# openapi.json
{
  "components": {
    "schemas": {
...
      "CapabilityResponse": {
        "properties": {
          "attributes": {
            "additionalProperties": true,
            "title": "Attributes",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "entity_id": {
            "format": "uuid",
            "title": "Entity Id",
            "type": "string"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "lifecycle": {
            "title": "Lifecycle",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "entity_id",
          "tenant_id",
          "name",
          "external_id",
          "lifecycle",
          "attributes",
          "created_at"
        ],
        "title": "CapabilityResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UpdateEntityRequest": {
        "description": "Bag of attribute updates applied bi-temporally; does not change entity_type or name directly.",
        "properties": {
          "updates": {
            "additionalProperties": true,
            "title": "Updates",
            "type": "object"
          },
          "valid_from": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          }
        },
        "required": [
          "updates"
        ],
        "title": "UpdateEntityRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "opaque",
        "description": "Paste an API token.\n\n- **Local development:** run `make dev-token` \u2014 it seeds a tenant + actor + role grants and prints a fresh token. No env-var setup required if the docker-compose stack is up.\n- **Production / break-glass:** mint with `python scripts/mint_token.py --tenant-id <uuid> --actor-id <uuid> --roles <role>`. Requires the tenant and actor to already exist.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
...
    "title": "registry",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/capabilities/{entity_id}": {
      "delete": {
        "description": "Soft-delete idempotency:\n- Row exists (active or already-invalidated) \u2192 204 No Content.\n- Row never existed \u2192 404 Not Found (service raises NotFoundError).",
        "operationId": "delete_capability_v1_capabilities__entity_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Capability",
        "tags": [
          "capabilities"
        ]
      },
      "get": {
        "description": "Return the full capability record.\n\nThe path segment accepts either a UUID or a slug-form name \u2014 they\nresolve to the same record. Slugs are case-insensitive against the\nstored `name` column.\n\nOptional `?as_of=` activates bi-temporal time-travel.\n\nOptional `?include=` adds bounded sub-resources to the response. Use\nthis to collapse \"fetch capability + components + facts + external\nIDs\" from four round-trips into one.\n\nOptional `?view=audit` returns the full bitemporal + tenant-id +\nsupersession audit shape. Default `view=default` omits those fields.",
        "operationId": "get_capability_v1_capabilities__entity_id__get",
        "parameters": [
          {
            "description": "Capability UUID or slug-form name (e.g. 'salt-design-system')",
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "description": "Capability UUID or slug-form name (e.g. 'salt-design-system')",
              "title": "Entity Id",
              "type": "string"
            }
          },
          {
            "description": "ISO-8601 UTC datetime for bi-temporal time-travel",
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO-8601 UTC datetime for bi-temporal time-travel",
              "title": "As Of"
            }
          },
          {
            "description": "Comma-separated list of sub-resources to expand. Known values: components, depends_on, external_ids, interface. Each expansion is capped at 200 items \u2014 `truncated: true` + a `next` URL signal overflow.",
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated list of sub-resources to expand. Known values: components, depends_on, external_ids, interface. Each expansion is capped at 200 items \u2014 `truncated: true` + a `next` URL signal overflow.",
              "title": "Include"
            }
          },
          {
            "description": "Response shape. `default` (UI-flavoured) is the standard minimal shape every endpoint returns. `audit` adds bitemporal columns (valid_from / valid_to / ingested_at / invalidated_at), tenant_id, and supersession metadata for audit / compliance consumers.",
            "in": "query",
            "name": "view",
            "required": false,
            "schema": {
              "default": "default",
              "description": "Response shape. `default` (UI-flavoured) is the standard minimal shape every endpoint returns. `audit` adds bitemporal columns (valid_from / valid_to / ingested_at / invalidated_at), tenant_id, and supersession metadata for audit / compliance consumers.",
              "title": "View",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated list of fact categories to include in the `facts` field (e.g. `release_note,overview`). Default: no filter \u2014 every category is returned. Use this to narrow a fat detail response, e.g. fetch Salt + just its release notes in one call.",
            "in": "query",
            "name": "facts_categories",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated list of fact categories to include in the `facts` field (e.g. `release_note,overview`). Default: no filter \u2014 every category is returned. Use this to narrow a fat detail response, e.g. fetch Salt + just its release notes in one call.",
              "title": "Facts Categories"
            }
          },
          {
            "description": "Cap the number of facts returned (applied after the category filter). Default: no cap. Useful when a capability has hundreds of facts and the UI only renders the top N.",
            "in": "query",
            "name": "facts_limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 500,
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cap the number of facts returned (applied after the category filter). Default: no cap. Useful when a capability has hundreds of facts and the UI only renders the top N.",
              "title": "Facts Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilityDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Capability",
        "tags": [
          "capabilities"
        ]
      },
      "patch": {
        "description": "Update mutable attributes on a capability.\n\nHonours the ``If-Match`` request header (advisory): if present and stale,\nreturns 412 Precondition Failed; if absent, logs a warning and accepts the\nwrite.  ETag is computed from the entity row before the write so a stale\nprecondition fails fast without executing the mutation.\n\nWhen there is no detail GET that a client can use to acquire the ETag,\nthe client may compute it from the list response or from a prior PATCH\nresponse body \u2014 but that is uncommon.  The recommended flow is:\nGET /v1/capabilities/{id} \u2192 ETag header \u2192 PATCH with If-Match.",
        "operationId": "patch_capability_v1_capabilities__entity_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEntityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Capability",
        "tags": [
          "capabilities"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}